feat: update teams header
